+++ src/config.ts
@@ -1,7 +1,7 @@
 import { ZKTypeConfig } from "./types";
 
 export const defaultCircuitArtifactGeneratorConfig: ZKTypeConfig = {
-  inputDir: "circuits",
+  circuitsASTPaths: [],
   outputArtifactsDir: "artifacts/circuits",
   outputTypesDir: "generated-types/circuits",
   clean: true,