+++ test/CircuitArtifactGenerator.test.ts
@@ -23,7 +23,13 @@ describe("Circuit Artifact Generation", function () {
   const inputDir = "test/fixture";
   const astDir = "test/cache/circuits-ast";
   const artifactGenerator = new CircuitArtifactGenerator({
-    inputDir: astDir,
+    circuitsASTPaths: [
+      "test/cache/circuits-ast/Basic.json",
+      "test/cache/circuits-ast/credentialAtomicQueryMTPV2OnChainVoting.json",
+      "test/cache/circuits-ast/lib/BasicInLib.json",
+      "test/cache/circuits-ast/auth/EMultiplier.json",
+      "test/cache/circuits-ast/auth/BasicInAuth.json",
+    ],
     outputArtifactsDir: defaultCircuitArtifactGeneratorConfig.outputArtifactsDir,
     clean: defaultCircuitArtifactGeneratorConfig.clean,
   });
@@ -53,49 +59,39 @@ describe("Circuit Artifact Generation", function () {
   });
 
   it("should throw an error if the compiler version field is missing", async function () {
-    fs.cpSync("test/mocks/InvalidCompilerVersion.json", `${astDir}/InvalidCompilerVersion.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(artifactGenerator.extractArtifact("test/mocks/InvalidCompilerVersion.json")).to.be.rejectedWith(
       "The compiler version is missing in the circuit AST: test/fixture/InvalidCompilerVersion.circom",
     );
   });
 
   it("should throw an error if the name in the initialization block is missing", async function () {
-    fs.cpSync("test/mocks/InvalidInitializationBlockName.json", `${astDir}/InvalidInitializationBlockName.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(
+      artifactGenerator.extractArtifact("test/mocks/InvalidInitializationBlockName.json"),
+    ).to.be.rejectedWith(
       "The initializations field of initialization block is missing or incomplete in the circuit AST: test/fixture/InvalidInitializationBlockName.circom",
     );
   });
 
   it("should throw an error if the main component is missing", async function () {
-    fs.cpSync("test/mocks/InvalidMainComponent.json", `${astDir}/InvalidMainComponent.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(artifactGenerator.extractArtifact("test/mocks/InvalidMainComponent.json")).to.be.rejectedWith(
       "The main component is missing or incomplete in the circuit AST: test/fixture/InvalidMainComponent.circom",
     );
   });
 
   it("should throw an error if the id of the main component is missing", async function () {
-    fs.cpSync("test/mocks/InvalidMainComponentId.json", `${astDir}/InvalidMainComponentId.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(artifactGenerator.extractArtifact("test/mocks/InvalidMainComponentId.json")).to.be.rejectedWith(
       "The main component id is missing in the circuit AST: test/fixture/InvalidMainComponentId.circom",
     );
   });
 
   it("should throw an error if the template block is missing", async function () {
-    fs.cpSync("test/mocks/InvalidTemplateBlock.json", `${astDir}/InvalidTemplateBlock.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(artifactGenerator.extractArtifact("test/mocks/InvalidTemplateBlock.json")).to.be.rejectedWith(
       "The template for the circuit Multiplier2 could not be found.",
     );
   });
 
   it("should throw an error if the xtype of the initialization block is missing", async function () {
-    fs.cpSync("test/mocks/InvalidXTypeField.json", `${astDir}/InvalidXTypeField.json`);
-
-    await expect(artifactGenerator.generateCircuitArtifacts()).to.be.rejectedWith(
+    await expect(artifactGenerator.extractArtifact("test/mocks/InvalidXTypeField.json")).to.be.rejectedWith(
       "The initialization block xtype is missing in the circuit AST: test/fixture/InvalidXTypeField.circom",
     );
   });