+++ test/CircuitTypesGenerator.test.ts
@@ -16,7 +16,15 @@ describe("Circuit Types Generation", function () {
 
   const astDir = "test/cache/circuits-ast";
 
-  const circuitTypesGenerator = new CircuitTypesGenerator({ inputDir: astDir });
+  const circuitTypesGenerator = new CircuitTypesGenerator({
+    circuitsASTPaths: [
+      "test/cache/circuits-ast/Basic.json",
+      "test/cache/circuits-ast/credentialAtomicQueryMTPV2OnChainVoting.json",
+      "test/cache/circuits-ast/lib/BasicInLib.json",
+      "test/cache/circuits-ast/auth/EMultiplier.json",
+      "test/cache/circuits-ast/auth/BasicInAuth.json",
+    ],
+  });
 
   function getPathToGeneratedType(generatedTypePath: string) {
     return path.join(projectRoot, circuitTypesGenerator.getOutputTypesDir(), generatedTypePath);