+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [TBD]
+ 
+- Define a list of helper functions. Validation per input in circuit, conversions?
+- Do general testing.
+
 ## [v0.2.0]
 
 - Resolved an issue where arrays might have multiple dimensions, such as `bigint[][][]`.
@@ -8,9 +13,10 @@
     - Added ability to specify where the circuit's AST files are stored
     - Added ability to specify where to put generated artifacts and types
 - Implemented Zkit wrapper generation for given circuits
-- Used ejs instead of ts factory to render the wrapper class
-- Added generation of the hardhat runtime extension file for circuits 
-- Added a helper function to return the circuit object from given name
+- Used EJS instead of TS factory to render the wrapper class
+- Added generation of the Hardhat runtime extension file for circuits
+- Added a helper function to return the circuit object from a given name
+- Changed the ArtifactGeneratorConfig to accept an array of paths to circuit ASTs.
 
 ## [v0.1.1]